+++ packages/core/src/lib/editor/index.ts
@@ -15,22 +15,56 @@ import {
   $createParagraphNode,
   $getSelection,
   $isRangeSelection,
+  $isNodeSelection,
+  $applyNodeReplacement,
+  getDOMSelection,
+  $getNearestNodeFromDOMNode,
+  $parseSerializedNode, 
+  DecoratorNode,
+
+
   FORMAT_TEXT_COMMAND,
+  COMMAND_PRIORITY_LOW,
+  SELECTION_CHANGE_COMMAND,
+  CLICK_COMMAND,
+  KEY_BACKSPACE_COMMAND,
+  KEY_DELETE_COMMAND,
+  
+
   // LEXICAL CORE TYPES
   type Klass,
   type LexicalNode,
   type LexicalNodeReplacement,
+  type BaseSelection,
+  type DOMConversionMap,
+  type DOMConversionOutput,
+  type DOMExportOutput,
+  type NodeKey,
+  type SerializedLexicalNode,
+  type Spread,
+
   ParagraphNode,
   TextNode,
   defineExtension,
   safeCast,
+  $getNodeByKey,
 
 } from "lexical";
 
 export {
   // LEXICAL COMMANDS
   FORMAT_TEXT_COMMAND,
+  SELECTION_CHANGE_COMMAND,
+  COMMAND_PRIORITY_LOW,
+  CLICK_COMMAND,
+  KEY_BACKSPACE_COMMAND,
+  KEY_DELETE_COMMAND,
+  // LEXICAL NODES
+
+
+
   // LEXICAL DEFAULT NODES
+  DecoratorNode,
   HeadingNode,
   ParagraphNode,
   TextNode,
@@ -54,9 +88,15 @@ export {
   $getSelectionStyleValueForProperty,
   $patchStyleText,
   $createHeadingNode,
-  defineExtension,
+  $getNodeByKey,
+  defineExtension, 
   safeCast,
+  getDOMSelection,
+  $isNodeSelection,
+  $applyNodeReplacement,
+  $getNearestNodeFromDOMNode,
+  $parseSerializedNode
 };
 
 // LEXICAL CORE TYPES
-export type { Klass, LexicalNode, LexicalNodeReplacement };
+export type { Klass, LexicalNode, LexicalNodeReplacement, BaseSelection, DOMConversionMap, DOMConversionOutput, DOMExportOutput, NodeKey, SerializedLexicalNode, Spread };