+++ packages/core/src/lib/index.ts
@@ -1,32 +1,50 @@
 // LEXICAL CORE TYPES
 export type {
-  Klass,
-  LexicalNode,
-  LexicalNodeReplacement,
+  Klass, LexicalNode, LexicalNodeReplacement, BaseSelection, DOMConversionMap, DOMConversionOutput, DOMExportOutput, NodeKey, SerializedLexicalNode, Spread
 } from "./editor";
 export {
-  $createHeadingNode,
+  // LEXICAL COMMANDS
+  FORMAT_TEXT_COMMAND,
+  SELECTION_CHANGE_COMMAND,
+  COMMAND_PRIORITY_LOW,
+  CLICK_COMMAND,
+  KEY_BACKSPACE_COMMAND,
+  KEY_DELETE_COMMAND,
+  // LEXICAL NODES
+
+
+
+  // LEXICAL DEFAULT NODES
+  DecoratorNode,
+  HeadingNode,
+  ParagraphNode,
+  TextNode,
+  QuoteNode,
+  ListNode,
+  ListItemNode,
+  LinkNode,
+  OverflowNode,
+  HashtagNode,
+  TableNode,
+  TableCellNode,
+  TableRowNode,
+  CodeNode,
+  CodeHighlightNode,
+  AutoLinkNode,
   // LEXICAL UTILS
   $createParagraphNode,
   $getSelection,
   $isRangeSelection,
   $setBlocksType,
-  AutoLinkNode,
-  CodeHighlightNode,
-  CodeNode,
-  // LEXICAL COMMANDS
-  FORMAT_TEXT_COMMAND,
-  HashtagNode,
-  // LEXICAL DEFAULT NODES
-  HeadingNode,
-  LinkNode,
-  ListItemNode,
-  ListNode,
-  OverflowNode,
-  ParagraphNode,
-  QuoteNode,
-  TableCellNode,
-  TableNode,
-  TableRowNode,
-  TextNode,
+  $getSelectionStyleValueForProperty,
+  $patchStyleText,
+  $createHeadingNode,
+  $getNodeByKey,
+  defineExtension,
+  safeCast,
+  getDOMSelection,
+  $isNodeSelection,
+  $applyNodeReplacement,
+  $getNearestNodeFromDOMNode,
+  $parseSerializedNode
 } from "./editor";