+++ packages/utils/src/index.ts
@@ -18,3 +18,7 @@ export {
   initializeDocumentState,
   validateEditorState,
 } from "./server/validation";
+
+export { $findMatchingParent, mergeRegister } from "./lexical/utils";
+export { getSelectedNode } from "./lexical/selected-node/index";
+export { getThemeSelector } from "./lexical/theme-selector/index";